Fix: 아티스트페이지 수정
