+++ client/src/icons/EditIcon.tsx
@@ -1,8 +1,8 @@
 const EditIcon = () => {
   return (
     <svg
-      width="24"
-      height="24"
+      width="20"
+      height="20"
       viewBox="0 0 24 24"
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
